Move phi mapping to separate struct.
