+++ atl-checker/src/algorithms/certain_zero/search_strategy/mod.rs
@@ -3,6 +3,7 @@ use crate::edg::{Edge, NegationEdge, Vertex};
 pub mod bfs;
 pub mod dependency_heuristic;
 pub mod dfs;
+mod linear_constrained_phi;
 mod linear_constraints;
 pub mod linear_optimize;
 pub mod linear_programming_search;
